refactor(config): move config/* files out of src/ root
CI check expects src/ root ≤ 2 files. alpha.39's config split put 5 new
files at src/ root (config-types.ts, config-load.ts, config-command.ts,
config-validate.ts, config-validate-ext.ts), failing CI.

Fix: move all into src/config/ subdirectory.
- src/config-types.ts → src/config/types.ts
- src/config-load.ts → src/config/load.ts
- src/config-command.ts → src/config/command.ts
- src/config-validate.ts → src/config/validate.ts
- src/config-validate-ext.ts → src/config/validate-ext.ts

src/config.ts (the barrel) stays at root and re-exports from ./config/*.
All external imports unchanged.

src/ root: 2 (cli.ts + config.ts). Tests: 725 pass / 0 fail.
